# backend/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    name: Delicious Team
  description: |-
    家庭菜谱管理系统 — 支持菜谱版本控制、百科搜索、翻译、导入导出
    所有接口前缀 /api/v1，单用户模式（无需认证）
  title: 人间烟火 (Delicious) API
  version: "2.0"
paths:
...
  /categories:
    get:
      description: 获取全部分类（字典表）
      produces:
      - application/json
      responses:
        "200":
          description: items
          schema:
            additionalProperties: true
            type: object
      summary: 分类列表
      tags:
      - 分类管理
    post:
      consumes:
      - application/json
      description: 新增一个分类名称
      parameters:
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: category
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 名称不能为空
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建分类
      tags:
      - 分类管理
  /categories/{id}:
    delete:
      description: 删除一个分类
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{}'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 分类不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除分类
      tags:
      - 分类管理
    put:
      consumes:
      - application/json
      description: 修改分类名称
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: category
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效 ID 或名称为空
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 修改分类
      tags:
      - 分类管理
  /dashboard/recipes/{id}/timeline:
    get:
      description: 获取菜谱完整时间线：基本信息 + 所有版本节点 + 当前版本标记
      parameters:
      - description: 菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: recipe, timeline[]
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 菜谱不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 菜谱时间线
      tags:
      - 数据面板
...
  /encyclopedia/{id}:
    get:
      description: 获取百科菜谱的完整信息，包括配料和制作步骤。浏览量自动 +1。支持翻译
      parameters:
      - description: 百科菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 返回语言 zh/en
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: recipe
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 菜谱不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 百科菜谱详情
      tags:
      - 百科菜谱
...
  /encyclopedia/category/{category}:
    get:
      description: 获取某个分类下的所有百科菜谱
      parameters:
      - description: 分类名称，如 川菜、西餐
        in: path
        name: category
        required: true
        type: string
      - description: 返回语言 zh/en
        in: query
        name: lang
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: items[], page_info
          schema:
            additionalProperties: true
            type: object
      summary: 按分类浏览百科
      tags:
      - 百科菜谱
  /encyclopedia/search:
    get:
      description: 搜索百科菜谱库。优先从外部 API（Spoonacular + MealDB）搜索，无结果或禁用时回退本地数据库。支持中英翻译
      parameters:
      - description: 搜索关键词（中英文均可）
        in: query
        name: keyword
        type: string
      - description: 分类筛选
        in: query
        name: category
        type: string
      - default: en
        description: 返回语言 zh/en
        in: query
        name: lang
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: items[], page_info
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 搜索百科菜谱
      tags:
      - 百科菜谱
...
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"